Simulate a dataset given model parameters.

Generates samples from a set of provided LATER model parameters, with the option to iteratively replace invalid samples (reaction times <= 0).

Usage

simulate_dataset(
  n,
  later_mu,
  later_sd,
  early_sd = NULL,
  seed = NA,
  allow_negative_times = FALSE
)

Arguments

n

Number of samples (trials)

later_mu

Mean of the later component.

later_sd

Standard deviation of the later component.

early_sd

Standard deviation of the early component, or NULL if there is no early component (the default).

seed

Seed for the random number generator

allow_negative_times

If FALSE (the default), any random samples that have negative response times are iteratively replaced such that all returned samples are positive. If TRUE, no such replacement is performed.

Value

Vector of response times (in seconds)

Examples

simulate_dataset(n = 100, later_mu = 5, later_sd = 1)
#>   [1] 0.3181881 0.1605521 0.2122705 0.1939785 0.2765102 0.1866865 0.1909570
#>   [8] 0.2307250 0.3220951 0.2146639 0.2055755 0.1896478 0.1951621 0.1940086
#>  [15] 0.1535375 0.1896763 0.1840485 0.2653777 0.2405911 0.1884785 0.2363080
#>  [22] 0.2154021 0.1722719 0.1982815 0.1814635 0.2302942 0.2745929 0.2480647
#>  [29] 0.1957557 0.1888618 0.1592636 0.2005592 0.2539864 0.2267716 0.1844435
#>  [36] 0.1417627 0.2327413 0.2495892 0.3756145 0.1967198 0.1717124 0.2054147
#>  [43] 0.2178967 0.2338260 0.1926590 0.2191031 0.3902393 0.1821361 0.2180623
#>  [50] 0.1541124 0.2605080 0.1874446 0.2801559 0.1507584 0.1950695 0.1729930
#>  [57] 0.2375536 0.2056625 0.2674502 0.2101465 0.2114932 0.2733438 0.2001578
#>  [64] 0.2314310 0.1487125 0.1701307 0.2066597 0.1806202 0.2763109 0.1701275
#>  [71] 0.1586348 0.2761430 0.1743375 0.2525240 0.1927489 0.2152814 0.1430782
#>  [78] 0.1991554 0.1864178 0.1927393 0.1630349 0.2279439 0.2579439 0.2345414
#>  [85] 0.1922815 0.2173665 0.2109049 0.2051371 0.1547568 0.1638172 0.2012478
#>  [92] 0.2628347 0.2119503 0.1863874 0.2049765 0.2680521 0.2001295 0.1877016
#>  [99] 0.1613625 0.2443900
simulate_dataset(n = 100, later_mu = 5, later_sd = 1, early_sd = 5)
#>   [1] 0.21751792 0.12198388 0.20999239 0.08948417 0.10880496 0.19029791
#>   [7] 0.20137210 0.16157372 0.16996214 0.22867048 0.16433555 0.20988253
#>  [13] 0.18306097 0.19582118 0.17112806 0.26042498 0.23253924 0.20825053
#>  [19] 0.17920393 0.18006877 0.19251792 0.15612947 0.21535333 0.10988556
#>  [25] 0.17477964 0.14360144 0.11433572 0.19767500 0.18610876 0.29048094
#>  [31] 0.20564780 0.21675407 0.16129103 0.22370545 0.16532450 0.30068211
#>  [37] 0.20519598 0.20233245 0.17390709 0.15256972 0.25691693 0.20484808
#>  [43] 0.18550608 0.15077674 0.29052551 0.18262795 0.17915548 0.29049350
#>  [49] 0.22232752 0.14760525 0.13049469 0.08532879 0.16390122 0.19656190
#>  [55] 0.25505293 0.25144823 0.14170014 0.17100263 0.18043449 0.15821250
#>  [61] 0.21091618 0.16390026 0.23543402 0.25911822 0.13586882 0.13195285
#>  [67] 0.16763946 0.20380078 0.13814439 0.15228739 0.22333549 0.19423470
#>  [73] 0.13799003 0.12939362 0.09920727 0.12548828 0.32087074 0.21393204
#>  [79] 0.16929819 0.25809578 0.23646230 0.09159899 0.17030354 0.16647863
#>  [85] 0.21398483 0.11767951 0.14660484 0.17493100 0.17255482 0.19447137
#>  [91] 0.13983553 0.09247451 0.19149097 0.15578723 0.07462491 0.25163179
#>  [97] 0.15006779 0.19415228 0.18122328 0.19377680